move prompt template to metadata
