+++ tests/transformers/test_transformers_prompt_templating.py
@@ -7,8 +7,11 @@ from transformers import pipeline
 import mlflow
 from mlflow.exceptions import MlflowException
 from mlflow.models.model import MLMODEL_FILE_NAME
-from mlflow.transformers import _SUPPORTED_PROMPT_TEMPLATING_TASK_TYPES
-from mlflow.utils.model_utils import _validate_prompt_template
+from mlflow.transformers import (
+    _PROMPT_TEMPLATE_KEY,
+    _SUPPORTED_PROMPT_TEMPLATING_TASK_TYPES,
+    _validate_prompt_template,
+)
 
 # session fixtures to prevent saving and loading a ~400mb model every time
 TEST_PROMPT_TEMPLATE = "Answer the following question like a pirate:\nQ: {prompt}\nA: "
@@ -73,9 +76,9 @@ def saved_transformers_model_path(tmp_path_factory, small_text_generation_model)
 )
 def test_prompt_validation_throws_on_invalid_templates(template):
     match = (
-        "Argument `prompt_template` should be a string with a single format arg, 'prompt'."
+        "Argument `prompt_template` must be a string with a single format arg, 'prompt'."
         if isinstance(template, str)
-        else "Argument `prompt_template` should be a string"
+        else "Argument `prompt_template` must be a string"
     )
     with pytest.raises(MlflowException, match=match):
         _validate_prompt_template(template)
@@ -100,7 +103,7 @@ def test_prompt_save_and_load(saved_transformers_model_path):
     with open(mlmodel_path) as f:
         mlmodel_dict = yaml.safe_load(f)
 
-    assert mlmodel_dict["prompt_template"] == TEST_PROMPT_TEMPLATE
+    assert mlmodel_dict["metadata"][_PROMPT_TEMPLATE_KEY] == TEST_PROMPT_TEMPLATE
 
     model = mlflow.pyfunc.load_model(saved_transformers_model_path)
     assert model._model_impl.prompt_template == TEST_PROMPT_TEMPLATE
